+++ ZetPhoneApp/DatabaseFiller/Station.cs
@@ -8,12 +8,12 @@ namespace DatabaseFiller
     public class Station
     {
         public int Id { get; set; }
-        public float Altitude { get; set; }
-        public float Longitude { get; set; }
+        public double Altitude { get; set; }
+        public double Longitude { get; set; }
         public string Name { get; set; }
         public string Direction { get; set; }
 
-        public Station(float altitude, float longitude, string name, string direction)
+        public Station(double altitude, double longitude, string name, string direction)
         {
             Altitude = altitude;
             Longitude = longitude;
